Move useEngineCanvas to spatial package
