+++ packages/client/src/pages/location/location.tsx
@@ -31,7 +31,7 @@ import '../../engine'
 
 import Debug from '@ir-engine/client-core/src/components/Debug'
 import { useEngineInjection } from '@ir-engine/client-core/src/components/World/EngineHooks'
-import { useEngineCanvas } from '@ir-engine/client-core/src/hooks/useEngineCanvas'
+import { useEngineCanvas } from '@ir-engine/spatial/src/renderer/functions/useEngineCanvas'
 import { useBrowserCheck } from '@ir-engine/client-core/src/hooks/useUnsupported'
 import LocationPage from '@ir-engine/client-core/src/world/Location'
 import { useSpatialEngine } from '@ir-engine/spatial/src/initializeEngine'
